cleaing up FSAT compressibilityFactor
